+++ ReferenceTests/src/tests/block_recipes.jl
@@ -56,6 +56,12 @@ BLOCK_UPDATES = let
         Legend => :bgcolor,
         # not settable
         Colorbar => :colormap,
+        Colorbar => :values,
+        Colorbar => :colorrange,
+        Colorbar => :limits,
+        Colorbar => :lowclip,
+        Colorbar => :highclip,
+        Colorbar => :scale,
         # avoid
         Axis3 => :zoommode, Toggle => :toggleduration,
     ]
@@ -168,7 +174,7 @@ BLOCK_UPDATES = let
     # includes a bunch of layoutobservables, Legend entry defaults, formatters,
     # dim_converts, ...
     # not counted: skipped hotkey updates, visible updates from true to true
-    @test length(skipped) == 93 # out of 719
+    @test length(skipped) == 90 # out of 719
 
     settings
 end